Cars are loading correctly inside fleet management

+++ src/main/java/io/github/grano22/carfleetapp/usermanagement/contract/AccountOwnerDetailsResponse.java
@@ -6,7 +6,7 @@ public record AccountOwnerDetailsResponse(
     String email,
     String phone,
     String birthDate,
-    Integer credits,
+    double credits,
     String status,
     String[] roles,
     String[] permissions,
