+++ src/main/java/io/github/grano22/carfleetapp/usermanagement/contract/UserDetailsView.java
@@ -6,7 +6,7 @@ public record UserDetailsView(
     String email,
     String phone,
     String birthDate,
-    Integer credits,
+    double credits,
     String status,
     String createdAt,
     String updatedAt
